voice-agent/requirements.txt
livekit-agents>=0.1.0
python-dotenv>=1.0.0
requests>=2.28.0
openai>=1.0.0
anthropic>=0.5.0
deepgram-sdk>=2.11.0
elevenlabs>=0.2.24
sentence-transformers>=2.2.2
pinecone-client>=2.2.2
uvicorn>=0.24.0
fastapi>=0.100.0
pydantic>=2.4.0
websockets>=11.0.3
numpy>=1.24.0
aiohttp>=3.8.5 